# Conflicts:
#	src/backend/ci/core/process/api-process/src/main/kotlin/com/tencent/devops/process/api/op/OpSubPipelineRefResource.kt
#	src/backend/ci/core/process/biz-process/src/main/kotlin/com/tencent/devops/process/service/webhook/PipelineBuildWebhookService.kt
#	src/backend/ci/ext/tencent/artifactory/api-experience-tencent/src/main/kotlin/com/tencent/devops/experience/api/app/AppExperienceSecurityResource.kt
#	src/backend/ci/ext/tencent/openapi/api-openapi-tencent/src/main/kotlin/com/tencent/devops/openapi/api/apigw/desktop/ApigwDeskTopResource.kt
#	src/backend/ci/ext/tencent/process/api-process-tencent/src/main/kotlin/com/tencent/devops/process/api/app/AppPipelineBuildTencentResource.kt

+++ src/backend/ci/core/auth/biz-auth/src/main/kotlin/com/tencent/devops/auth/service/ApiAccessTokenService.kt
@@ -51,7 +51,7 @@ class ApiAccessTokenService @Autowired constructor(
     val dslContext: DSLContext
 ) {
     @Value("\${auth.accessToken.expirationTime:#{null}}")
-    private val expirationTime: Int? = null
+    private val expirationTime: Double? = null
 
     @Value("\${auth.accessToken.secret:#{null}}")
     private val secret: String? = null
@@ -94,7 +94,7 @@ class ApiAccessTokenService @Autowired constructor(
         }
         val tokenInfo = TokenInfo(
             userId = userDetails,
-            expirationTime = System.currentTimeMillis() + (expirationTime ?: EXPIRE_TIME_MILLS),
+            expirationTime = System.currentTimeMillis() + (expirationTime ?: EXPIRE_TIME_MILLS).toLong(),
             accessToken = null
         )
         tokenInfo.accessToken = try {
@@ -146,7 +146,7 @@ class ApiAccessTokenService @Autowired constructor(
     }
 
     companion object {
-        private const val EXPIRE_TIME_MILLS: Int = 14400000
+        private const val EXPIRE_TIME_MILLS: Double = 14400000.0
         private val logger = LoggerFactory.getLogger(ApiAccessTokenService::class.java)
     }
 }
